requirements.txt
langchain==0.1.9
langchain_openai==0.0.7
openai==1.12.0
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.31.1
